automatically backed up by learn

+++ guessing_cli.rb
@@ -5,7 +5,7 @@ def run_guessing_game
   random_num = 1 + rand(6)
   while user_inputs != "exit"
 
-    if user_inputs.to_f == random_num
+    if user_inputs.to_i == random_num
       puts "You guessed the correct number!"
       break
     else
